[Patchset 1] Remove warning suppressions
WIP

+++ include/perfetto/ftrace_reader/ftrace_controller.h
@@ -45,7 +45,7 @@ using Inode = decltype(stat::st_ino);
 struct FtraceMetadata {
   FtraceMetadata();
 
-  size_t overwrite_count;
+  uint32_t overwrite_count;
   BlockDeviceID last_seen_device_id = 0;
 #if PERFETTO_DCHECK_IS_ON()
   bool seen_device_id = false;
